Move the architecture loop out of scripts
All the script already take a 'distro' argument, and the for loop on
distros is done in the gitlab ci. It seems sensible to do the same with
architectures, if only for consistency.

A side-effect is that I have to temporarily disable the Docker Hub
manifest thing, as it doesn't fit well. I'll need to investigate that
and re-enable it later.