Rework how release version is passed form build-rootfs to docker-build
I broke it in 29c33f52f15a18e76bab9a107458b516518ea78b

Now I bring it back, but instead of naming the file 'release.version',
I prefer to name it "$ARCH.$DISTRO.release.version", to couple it more
tightly with "$ARCH.$DISTRO.tar.xz" (as this file belong to this
particular archive).

Note that in gitlab ci's 'push-docker-images', RELEASE was set but was
never used, as far as I can see. So let's remove it.