Rework DOCKER_BUILDKIT and DOCKER_CLI_EXPERIMENTAL usage
No need to have those variables set globally.

For 'docker buildx', we need DOCKER_CLI_EXPERIMENTAL. However no need
for DOCKER_BUILDKIT=1:

    Buildx builds using the BuildKit engine and does not require
    DOCKER_BUILDKIT=1 environment variable to start the builds.

For 'docker build', no need for DOCKER_CLI_EXPERIMENTAL. However we want
to enable the buildkit engine, so we set 'DOCKER_BUILDKIT=1', so that
buildkit is used, and we have the '--platform' argument.

At this point, I don't know why we use 2 different methods, depending on
whether we are in GitLab CI (CI_JOB_TOKEN) or not.

BTW, this commit also adds a seemingly missing --platform argument.