Re-instate docker push for GitLab CI
My bad, only now I understand that different stages are picked up by
different runners, and the most direct way to keep docker images between
each stages is to actually push them to a registry.

Feels a bit lame to push the image out there before actually running the
test stage, as we might push a broken image and realize only later. But
that's how things work. The bad is somewhat mitigated by NOT tagging the
image as 'latest' at this point. We do that only after the test passed.

EDIT: Only now I seem to understand that it does not really matter. The
images that are pushed on the GitLab registry, despite being *public*,
are not *published*. They are there only as part of our CI workflow, but
this is not where we direct users for Kali Docker images. So it does not
really matter what images get pushed there.