docker-build: Change version from 'Y-m-d' to 'Y.m.d'
This is because now, the image tag is set to 'VERSION-ARCH'. In case
we'd need to parse that, better make sure that '-' is the separator,
and there's no '-' in the version.

So 2021-03-25-amd64 becomes 2021.03.25-amd64.

This change has no impact on images pushed to the Docker Hub, as only
the latest tag is pushed there.

To say it again. This change impacts the next images that will be pushed
to GitLab. The last tags that were pushed are `2021-03-19-armhf`,
`2021-03-19-arm64` and so on.  The next tags that will be pushed will be
something like `2021.03.26-amd64`.