Completely rework docker push stage
Now we push the manifests, as needed for multi-arch.

We follow the following references:
- https://www.docker.com/blog/multi-arch-build-and-images-the-simple-way/
- https://docs.gitlab.com/ee/user/packages/container_registry/#blob-unknown-to-registry-error-when-pushing-a-manifest-list

Let's be a bit more detailed.

At the moment in Docker Hub, tags are `latest`, `amd64`, `armhf` and
`arm64`. The tag `latest` points to `amd64`.

With this commit, the tag `latest` will now be a manifest that refers to
all of those images.

At the moment in GitLab, tags are `2021-03-19-amd64`,
`2021-03-19-arm64`, `2021-03-19-armhf` and `latest`. `latest` just
points to `2021-03-19-amd64`.

With this commit, we will have two manifests. `2021-03-19`, that will
refer to the 3 images, and also `latest` to refer to the 3 images.