Drop buildx, not needed
What is REALLY needed for multi-arch images is buildkit, which is more
or less a backend for 'docker build', as I can understand.

There's no need to install anything, this is already part of docker for
a while. All is needed is DOCKER_BUILDKIT=1 in the environment of
'docker build'.

buildx is an experimental command that is meant to simplify building
multi-arch images, among other things, by doing everything in one round,
including creating a manifest and pushing.

We could very well use it, that would simplify things a bit, but not
that much. Since we push to TWO different docker registries anyway
(gitlab and docker hub), we have to do the manifest part manually at
some point, in order to push it to the Docker Hub.

I don't think that we can rename / retag a manifest, and that it would
also automagically retag all the images that it contains. Although I'm
not sure, since after all, images are described by sha sums, and it's
exactly the same images that are on both Docker Hub and GitLab.

Anyway. Let's drop buildx for now and stick with plain old 'docker
build'.