build-rootfs: Workaround debootstrap's occasional failure
The script build-rootfs is called many times to build various kali
distros, and various architectures for each of those distros.

It happens, from time to time, that the debootstrap call fails:

    Building rootfs kali-last-snapshot/arm64
    E: Unable to execute target architecture

This particular error message comes from the tool arch-test (called from
within debootstrap), when it tries and fails to execute a test binary in
a foreign architecture.

The reason for this failure is still mysterious. It's not easily
reproduced. And when it fails to build the kali-last-snapshot distro,
it means that it succeeded building kali-dev and kali-rolling before.
Nothing changed in the build environment in between.

So this commit simply wraps debootstrap in a retry loop, hoping that
it's enough. In case of failure, we dump some information about the
environment, in case it can help to understant what's wrong.

Fixes: #37