docker-push-manifest: Delete local manifests after pushing it
Not needed when running on GitLab CI, but definitely needed for local
runs. Otherwise we'll leave manifests around, and in the next run, new
images will be added alongside old ones in the existing manifests,
instead of creating new manifests.

Another approach could be to create the manifest with 'create' instead
of always running 'create --amend', in docker-push.sh. I *guess* it
would delete an existing manifest, if any? But that would be so awkward,
as it's not possible to create an empty manifest. So when we loop on
architectures, we need a special case for the first iter of the loop,
where we do 'manifest create', then in the next iterations we'll do
'manifest create --amend'... So awkward.

Local manifests are a bit of a mysterious thing, because there's simply
no way to list them via the cli, so it's very easy to forget that it
exists: https://stackoverflow.com/a/67406040/776208