Image-based build
At a first glance, it might seem that this commit is just renaming
'distro' to 'image'. But there's a bit more to it.

Up to now, we've been building *one container image per distro*. The
argument for each script was the distro, and then for each distro
corresponded an image.  The image name was equal to the distro name, the
only special case being the distro 'kali-last-snapshot', mapped to the
image 'kali'.

With this commit, we reverse this logic: we build different images, and
it just happens that to each image corresponds a distro. The argument
for each script is now the image, and the script deduces the distro from
the image name, if needed. Actually, only the script build-rootfs.sh
needs to know the distro, in order to debootstrap. All the other scripts
don't care, they just deal with images.

Why this change? First, it makes things simpler. We don't need to have
an IMAGE variable in the conf files anymore, since IMAGE is now part of
the conf filename.

This commit also prepares for the future. We might want to build some
images for special purpose.  Imagine kali-with-basic-tools for an image
that comes with some tools pre-installed, for example. This was simply
not possible when the scripts were "distro-based". So making the scripts
"image-based" instead will make that easier.

Incidentelly, with this commit, the image 'kali' ends up being renamed
in 'kali-last-snapshot'. That's ok, we're about to address that in the
next commit.