Switch to podman (work in progress)
Use podman in place of docker. At this point, the docker-push.sh and
docker-push-manifest.sh scripts are probably not functional, there's
nothing more than s/docker/podman/ in there, and it's not been tested.

Other than that, there's a few things interesting.

It looks simpler in .gitlb-ci.yml, since podman is not a daemon, so it
doesn't require a separate service running in another container, as it
was done before.

The script build.sh becomes simpler as well, since we don't need root
permissions anymore, except for debootstrap.

The build command changed a bit, also it didn't have to. A while ago,
podman didn't support --platform, so we had to use --arch (cf. [1]
below). Nowadays it does support --platform though, so I guess a simple
s/docker/podman/ could have been enough.

In any case, using --arch is easier, as it accepts Debian-style
architecures.

At last, the 'manifest' command seems quite different in podman, it has
much more sub-commands, much more arguments, and the argument --amend
that we use is... not there yet. But it was just added in the master
branch (a month ago) and will be released pretty soon, I guess. Cf [2].

This last point will be addressed in the next commit.

----

[1]: Cf. and old `podman build(1)` manual page:

    --arch=arch Set the ARCH of the image to the provided value
        instead of the architecture of the host.

    --platform="Linux" This option has no effect on the build. Other
        container engines use this option to control the execution
        platform for the build (e.g., Windows, Linux) which is not
        required for Buildah as it supports only Linux.

[2]: https://github.com/containers/podman/issues/15346