Complete rework of docker-push{,manifest}.sh, fix podman usage
Merge those two scripts under the name docker-publish.sh, and try hard
to simplify the code and make it readable. What we do in those scripts
is nothing complicated, but for some reasons the scripts themselves are
super hard to work with and are headache each time I have to poke it.

Let's try harder to make it simple.

Apart from that, this commit completes the previous one, and makes
everythng functional again. The 'podman manifest' commands have a
somewhat different syntax that their 'docker manifest' counterparts.

For reference:
- https://podman.io/blogs/2021/10/11/multiarch.html