Misc improvements in docker-build-extra.sh
We need to know CI_REGISTRY_IMAGE and TAG, which is done by sourcing the
conf file. That's what we doe everywhere else, why not in this script?

Then it turns out that the conf file is the same as the conf file from
the base image, so we just need to copy it, rather than rebuild it.