Drop pixz, back to gzip
It seems that the point of pixz is to compress faster using multiple
processors. However we don't really care in GitLab CI case, and I'm
not even sure we have access to multiple processors.

As for building locally, this is probably more of an annoyance than
anything else, as it will take too much CPU and slow down my machine
while I'm trying to do other things.

So let's get back to standard, fast gzip compression. Sure it doesn't
compress as much as xz, but we don't care.