Push images to the Docker Hub in the v2s2 format
While looking at the images in the Docker Hub (in the admin view), eg.
<https://hub.docker.com/repository/docker/kalilinux/kali-rolling>, I
noticed that details from the image are missing (digest, os/arch, ...).

So, to say it again: in the web page above, the details of the manifest
are good, but the details for each individual image are missing. I
believe it's because we let podman push the images in the default
format, which must be oci. So let's push using the format v2s2.

This is related to commit 5cc12db31b08db845cd0617289d660d9e9c8422d,
where I pushed the *manifest* in the v2s2 format.

Note that it probably doesn't matter for image consumers, it's really
just about fixing how it looks like in the Docker Hub admin page, and
also in case it fixes more things here and there. I think that, as long
as the Kali images are hosted on the Docker Hub, we probably want to
keep using the v2s2 format.

Also, note that it didn't seem to have any impact in the GitLab Registry
web interface. As far as I can see, there's just no details on the
images.